+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "export-dir2file" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.1.3] - Jul 5, 2024
+
+## [0.1.2] - Jul 5, 2024
+
+- `description` property in `exportconfig.json` is not could be string or object with `main` and `activeTabs` keys descriptions for export files and active tabs commands
+- fix of description not exported on some commands
+
 ## [0.1.0] - Jul 2, 2024
 
 - bug fixes